Methoden zum Setzten/Prüfen des Dirty Flags in eigene Klasse auslagern
closes #1272
