+++ bundles/aero.minova.rcp.rcp/src/aero/minova/rcp/rcp/parts/WFCDetailPart.java
@@ -26,13 +26,10 @@ import org.eclipse.e4.core.di.annotations.Optional;
 import org.eclipse.e4.core.di.extensions.Preference;
 import org.eclipse.e4.core.services.translation.TranslationService;
 import org.eclipse.e4.ui.di.PersistState;
-import org.eclipse.e4.ui.di.UIEventTopic;
 import org.eclipse.e4.ui.model.application.MApplication;
 import org.eclipse.e4.ui.model.application.commands.MCommand;
 import org.eclipse.e4.ui.model.application.commands.MParameter;
-import org.eclipse.e4.ui.model.application.ui.advanced.MPerspective;
 import org.eclipse.e4.ui.model.application.ui.basic.MPart;
-import org.eclipse.e4.ui.model.application.ui.basic.MTrimBar;
 import org.eclipse.e4.ui.model.application.ui.basic.MWindow;
 import org.eclipse.e4.ui.model.application.ui.menu.MHandledItem;
 import org.eclipse.e4.ui.model.application.ui.menu.MHandledMenuItem;
@@ -94,10 +91,6 @@ import aero.minova.rcp.form.setup.util.XBSUtil;
 import aero.minova.rcp.form.setup.xbs.Node;
 import aero.minova.rcp.form.setup.xbs.Preferences;
 import aero.minova.rcp.model.Column;
-import aero.minova.rcp.model.event.GridChangeEvent;
-import aero.minova.rcp.model.event.GridChangeListener;
-import aero.minova.rcp.model.event.ValueChangeEvent;
-import aero.minova.rcp.model.event.ValueChangeListener;
 import aero.minova.rcp.model.form.MBooleanField;
 import aero.minova.rcp.model.form.MButton;
 import aero.minova.rcp.model.form.MDateTimeField;
@@ -130,13 +123,14 @@ import aero.minova.rcp.rcp.fields.RadioField;
 import aero.minova.rcp.rcp.fields.ShortDateField;
 import aero.minova.rcp.rcp.fields.ShortTimeField;
 import aero.minova.rcp.rcp.fields.TextField;
+import aero.minova.rcp.rcp.util.DirtyFlagUtil;
 import aero.minova.rcp.rcp.util.TabUtil;
 import aero.minova.rcp.rcp.util.TranslateUtil;
 import aero.minova.rcp.rcp.util.WFCDetailCASRequestsUtil;
 import aero.minova.rcp.rcp.widgets.SectionGrid;
 
 @SuppressWarnings("restriction")
-public class WFCDetailPart extends WFCFormPart implements ValueChangeListener, GridChangeListener {
+public class WFCDetailPart extends WFCFormPart {
 
 	@Inject
 	@Preference(nodePath = ApplicationPreferences.PREFERENCES_NODE, value = ApplicationPreferences.TIMEZONE)
@@ -158,8 +152,6 @@ public class WFCDetailPart extends WFCFormPart implements ValueChangeListener, G
 
 	private MDetail mDetail = new MDetail();
 
-	private boolean dirtyFlag;
-
 	@Inject
 	private MPart mPart;
 
@@ -174,6 +166,7 @@ public class WFCDetailPart extends WFCFormPart implements ValueChangeListener, G
 	private EHandlerService handlerService;
 	private LocalResourceManager resManager;
 	private WFCDetailCASRequestsUtil casRequestsUtil;
+	private DirtyFlagUtil dirtyFlagUtil;
 
 	private IEclipseContext appContext;
 
@@ -212,10 +205,15 @@ public class WFCDetailPart extends WFCFormPart implements ValueChangeListener, G
 			return;
 		}
 
+		// DiryFlagUtil erstellen und in Kontext setzten
+		dirtyFlagUtil = ContextInjectionFactory.make(DirtyFlagUtil.class, mPart.getContext());
+		mPerspective.getContext().set(DirtyFlagUtil.class, dirtyFlagUtil);
+
 		layoutForm(parent);
 
 		mDetail.setDetailAccessor(new DetailAccessor(mDetail));
 		ContextInjectionFactory.inject(mDetail.getDetailAccessor(), mPerspective.getContext()); // In Context, damit Injection verfügbar ist
+		mPerspective.getContext().set(MDetail.class, mDetail); // MDetail per Injection ermöglichen
 
 		mDetail.setClearAfterSave(form.getDetail().isClearAfterSave());
 
@@ -1019,7 +1017,7 @@ public class WFCDetailPart extends WFCFormPart implements ValueChangeListener, G
 		String fieldName = suffix + field.getName();
 		try {
 			MField f = ModelToViewModel.convert(field, locale);
-			f.addValueChangeListener(this);
+			f.addValueChangeListener(dirtyFlagUtil);
 			f.setName(fieldName);
 
 			getDetail().putField(f);
@@ -1039,7 +1037,7 @@ public class WFCDetailPart extends WFCFormPart implements ValueChangeListener, G
 	private void createGrid(Composite composite, MSection mSection, MinovaSection section, IEclipseContext context, Object fieldOrGrid) {
 		SectionGrid sg = new SectionGrid(composite, section, (Grid) fieldOrGrid, mDetail);
 		MGrid mGrid = createMGrid((Grid) fieldOrGrid, mSection);
-		mGrid.addGridChangeListener(this);
+		mGrid.addGridChangeListener(dirtyFlagUtil);
 		GridAccessor gA = new GridAccessor(mGrid);
 		gA.setSectionGrid(sg);
 		mGrid.setGridAccessor(gA);
@@ -1123,85 +1121,6 @@ public class WFCDetailPart extends WFCFormPart implements ValueChangeListener, G
 
 	public WFCDetailCASRequestsUtil getRequestUtil() {
 		return casRequestsUtil;
-	}
-
-	private void setDirtyFlag(boolean dirtyFlag) {
-		this.dirtyFlag = dirtyFlag;
-
-		mPart.setDirty(dirtyFlag);
-		@SuppressWarnings("unchecked")
-		List<MPerspective> pList = (List<MPerspective>) appContext.get(Constants.DIRTY_PERSPECTIVES);
-
-		if (dirtyFlag) {
-			if (pList == null) {
-				pList = new ArrayList<>();
-				appContext.set(Constants.DIRTY_PERSPECTIVES, pList);
-			}
-			if (!pList.contains(mPerspective)) {
-				pList.add(mPerspective);
-				refreshToolbar();
-			}
-		} else {
-			if (pList != null) {
-				pList.remove(mPerspective);
-				refreshToolbar();
-			}
-		}
-	}
-
-	/**
-	 * True wenn es Änderungen gab, False ansonsten
-	 *
-	 * @return
-	 */
-	public boolean getDirtyFlag() {
-		return dirtyFlag;
-	}
-
-	public void refreshToolbar() {
-		List<MTrimBar> findElements = eModelService.findElements(mwindow, "aero.minova.rcp.rcp.trimbar.0", MTrimBar.class);
-		MTrimBar tBar = findElements.get(0);
-		Composite c = (Composite) (tBar.getChildren().get(0)).getWidget();
-		if (c == null) {
-			return;
-		}
-		ToolBar tb = (ToolBar) c.getChildren()[0];
-
-		String perspectiveLabel = translationService.translate(mPerspective.getLabel(), null);
-		for (ToolItem item : tb.getItems()) {
-			if (item.getText().replace("*", "").equals(perspectiveLabel)) {
-				item.setText((dirtyFlag ? "*" : "") + perspectiveLabel);
-			}
-		}
-		tb.requestLayout();
-	}
-
-	@Override
-	public void gridChange(GridChangeEvent evt) {
-		checkDirtyFlag();
-	}
-
-	@Override
-	public void valueChange(ValueChangeEvent evt) {
-		checkDirtyFlag();
-	}
-
-	@Inject
-	@Optional
-	public void checkDirtyFromBroker(@UIEventTopic(Constants.BROKER_CHECKDIRTY) String message) {
-		MPerspective activePerspective = eModelService.getActivePerspective(mwindow);
-		if (activePerspective.equals(mPerspective)) {
-			checkDirtyFlag();
-		}
-	}
-
-	private void checkDirtyFlag() {
-		if (casRequestsUtil != null) {
-			boolean setDirty = casRequestsUtil.checkDirty();
-			if (this.dirtyFlag != setDirty) {
-				setDirtyFlag(setDirty);
-			}
-		}
 	}
 
 	@PersistState
